+++ test-sql/sqllogic/runtest2.js
@@ -10,41 +10,43 @@ var md5 = require('blueimp-md5').md5;
 // List of tests
 var filenames = [
 //	'./test/select1.test',
- 	'./test/select2.test',
+// 	'./test/select2.test',
 //	'./test/select3.test',
 // 	'./test/select4.test',
-//	'./test/select5.test',
-/*	'./test/evidence/in1.test',
-	'./test/evidence/in2.test',
- 	'./test/evidence/slt_lang_aggfunc.test',
- 	'./test/evidence/slt_lang_createtrigger.test',
- 	'./test/evidence/slt_lang_createview.test',
- 	'./test/evidence/slt_lang_dropindex.test',
- 	'./test/evidence/slt_lang_droptable.test',
- 	'./test/evidence/slt_lang_droptrigger.test',
- 	'./test/evidence/slt_lang_dropview.test',
- 	'./test/evidence/slt_lang_reindex.test',
- 	'./test/evidence/slt_lang_replace.test',
- 	'./test/evidence/slt_lang_update.test',
-	 './test/index/between/1/slt_good_0.test',
-	 './test/index/commute/10/slt_good_0.test',
- 	 './test/index/delete/10/slt_good_0.test',
- 	 './test/index/in/10/slt_good_0.test',
- 	 './test/index/orderby/10/slt_good_0.test',
- 	 './test/index/orderby_nosort/10/slt_good_0.test',
- 	 './test/index/random/10/slt_good_0.test',
- 	 './test/index/view/10/slt_good_1.test',
+	'./test/select5.test',
+//	'./test/evidence/in1.test',
+//	'./test/evidence/in2.test',
+// 	'./test/evidence/slt_lang_aggfunc.test',
+// 	'./test/evidence/slt_lang_createtrigger.test',
+// 	'./test/evidence/slt_lang_createview.test',
+// 	'./test/evidence/slt_lang_dropindex.test',
+// 	'./test/evidence/slt_lang_droptable.test',
+// 	'./test/evidence/slt_lang_droptrigger.test',
+// 	'./test/evidence/slt_lang_dropview.test',
+// 	'./test/evidence/slt_lang_reindex.test',
+// 	'./test/evidence/slt_lang_replace.test',
+// 	'./test/evidence/slt_lang_update.test',
+//	 './test/index/between/1/slt_good_0.test',
+//	 './test/index/commute/10/slt_good_0.test',
+ //	 './test/index/delete/1/slt_good_0.test',
+ //	 './test/index/in/10/slt_good_0.test',
+ //	 './test/index/orderby/10/slt_good_0.test',
+ //	 './test/index/orderby_nosort/10/slt_good_0.test',
+ //	 './test/index/random/10/slt_good_0.test',
+ //	 './test/index/view/10/slt_good_1.test',
 
- 	 './test/random/aggregates/slt_good_1.test',
- 	 './test/random/expr/slt_good_1.test',
- 	 './test/random/groupby/slt_good_1.test',
- 	 './test/random/select/slt_good_1.test'
-*/
+// 	 './test/random/aggregates/slt_good_0.test',
+// 	 './test/random/expr/slt_good_0.test',
+// 	 './test/random/groupby/slt_good_0.test',
+// 	 './test/random/select/slt_good_0.test'
+
 
 ];
 
 
-var limit = 1000000; /*1000000*/
+var limit = 200000; /*1000000*/
+var errlimit = 2;
+var nerrors = 0;
 //var mode = 'PostgreSQL';		// Let say we are a la Oracle :)
 var mode = 'oracle';		// Let say we are a la Oracle :)
 
@@ -109,8 +111,8 @@ function test(filename, show) {
 
 	for(var i = 0;i < Math.min(limit,f.length);i++) {
 //		if(i>3100 && i<29209) continue;
-		if(i%100 == 0) process.stdout.write('.');
-//		process.stdout.write(i+',');
+//		if(i%100 == 0) process.stdout.write('.');
+		process.stdout.write(i+',');
 
 		var line = f[i].trim().split('#')[0];
 		var w = line.split(' '); 
@@ -177,6 +179,7 @@ function test(filename, show) {
 
 //			console.log(131,qtype);
 			if(ast.statements[0].joins && ast.statements[0].joins.length > 3) { // We skip queries with joining more than 3 tables
+				process.write('0');
 				continue;
 			}
 
@@ -211,8 +214,14 @@ function test(filename, show) {
 					    var s = res.map(function(d){
 					    	var s1 = '';
 					    	for(var j=0;j<rs.columns.length;j++) {
-					    		s1 += d[rs.columns[j].columnid]+'\n';
+					    		if((typeof d[rs.columns[j].columnid] == 'undefined')
+					    		|| isNaN(d[rs.columns[j].columnid])){
+						    		s1 += 'NULL\n';
+					    		} else {
+						    		s1 += d[rs.columns[j].columnid]+'\n';
+					    		}
 					    	}
+					    //	console.log(s1);
 					    	return s1;
 					    }).join('');
     					rhash = md5(s);
@@ -235,6 +244,12 @@ function test(filename, show) {
 								} else {
 									passed = 'not passed';
 								}
+							} else {
+								if((expect[0]||0) == (res[0][rs.columns[0].columnid]||0)) {
+									passed = 'passed';
+								} else {
+									passed = 'not passed';
+								}
 							}				
 						} else {
 							var resvals = [];
@@ -243,9 +258,13 @@ function test(filename, show) {
 						    		resvals.push(d[rs.columns[j].columnid]);
 						    	}
 						    });
-
+						    passed = 'passed';
 						    for(var j=0;j<explen;j++) {
-						    	if(resvals[j] != expect[j]) {
+//						    	console.log(((expect[j] == 'NULL') && ((resvals[j]||0) == 0)));
+						    	if(((resvals[j]||0) == (expect[j]||0))
+									|| ((expect[j] == 'NULL') && ((resvals[j]||0) == 0))){
+//						    		console.log('ok');
+						    	} else {
 									passed = 'not passed';
 									break;
 						    	}
@@ -277,7 +296,7 @@ function test(filename, show) {
 				}
 			}
 //			if(passed != 'passed' && show) console.log(passed,i,sql,res,expect,reason);
-			if(passed != 'passed' && show) console.log(264,i,expect,reason);
+			if(passed != 'passed' && show) console.log('#',i,expect,reason);
 			if(passed == 'passed') npassed++;
 			else if(passed == 'unclear') upassed++;
 			continue;
@@ -323,7 +342,7 @@ function test(filename, show) {
 			}
 
 //			if(!passed && show) console.log(i,sql,res,expect,reason);
-			if(!passed && show) console.log(i,expect,reason);
+			if(!passed && show) console.log('#',i,expect,reason);
 			if(passed) npassed++;
 
 			continue;